app/core/middleware.py
"""Performance monitoring middleware."""
import time
from typing import Callable

from fastapi import Request, Response
from starlette.middleware.base import BaseHTTPMiddleware

from app.core.logging import get_logger
from app.utils.performance import metrics

logger = get_logger(__name__)


class PerformanceMiddleware(BaseHTTPMiddleware):
    """Middleware to track request performance."""

    async def dispatch(self, request: Request, call_next: Callable) -> Response:
        """
        Process request and track performance.
        
        Args:
            request: Incoming request
            call_next: Next middleware/handler
            
        Returns:
            Response
        """
        start_time = time.time()
        
        # Track request
        path = request.url.path
        method = request.method
        
        try:
            response = await call_next(request)
            duration_ms = (time.time() - start_time) * 1000
            
            # Log slow requests
            if duration_ms > 2000:  # 2 seconds
                logger.warning(
                    f"Slow request: {method} {path} took {duration_ms:.2f}ms"
                )
            
            # Add timing header
            response.headers["X-Process-Time"] = f"{duration_ms:.2f}ms"
            
            return response
            
        except Exception as e:
            duration_ms = (time.time() - start_time) * 1000
            logger.error(
                f"Request failed: {method} {path} after {duration_ms:.2f}ms - {e}"
            )
            raise
