requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.6
gunicorn==21.2.0

# Face Recognition & AI
deepface==0.0.92
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.24.3
tensorflow==2.15.0
tf-keras==2.15.0

# Utilities
python-json-logger==2.0.7
redis==5.0.1
requests==2.31.0

# Performance Optimization
imagehash==4.3.1
orjson==3.9.15
prometheus-client==0.19.0
